+++ backend/src/main/java/com/pobluesky/backend/domain/answer/controller/AnswerController.java
@@ -1,0 +1,28 @@
+package com.pobluesky.backend.domain.answer.controller;
+
+import com.pobluesky.backend.domain.answer.dto.request.AnswerCreateRequestDTO;
+import com.pobluesky.backend.domain.answer.dto.response.AnswerResponseDTO;
+import com.pobluesky.backend.domain.answer.service.AnswerService;
+import com.pobluesky.backend.global.util.ResponseFactory;
+import com.pobluesky.backend.global.util.model.JsonResult;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/answer")
+public class AnswerController {
+    private final AnswerService answerService;
+
+    @PostMapping("/answer/{inquiryId}")
+    public ResponseEntity<JsonResult> createAnswer(
+        @PathVariable Long inquiryId,
+        @RequestBody AnswerCreateRequestDTO answerCreateRequestDTO) {
+        AnswerResponseDTO response = answerService.createAnswer(inquiryId, answerCreateRequestDTO);
+        return ResponseEntity.status(HttpStatus.OK)
+
+            .body(ResponseFactory.getSuccessJsonResult(response));
+    }
+}
