+++ backend/src/main/java/com/pobluesky/backend/domain/answer/dto/request/AnswerCreateRequestDTO.java
@@ -1,14 +1,16 @@
-package com.pobluesky.backend.domain.voc.dto.request;
+package com.pobluesky.backend.domain.answer.dto.request;
 
-import com.pobluesky.backend.domain.voc.entity.Answer;
 import com.pobluesky.backend.domain.inquiry.entity.Inquiry;
+import com.pobluesky.backend.domain.answer.entity.Answer;
 
-public record VocAnswerCreateRequestDTO(
+public record AnswerCreateRequestDTO(
+    String answerTitle,
     String answerContents
 ) {
     public Answer toAnswerEntity(Inquiry inquiry) {
         return Answer.builder()
             .inquiry(inquiry)
+            .answerTitle(answerTitle)
             .answerContents(answerContents)
             .build();
     }
