+++ backend/src/main/java/com/pobluesky/backend/domain/answer/repository/AnswerRepository.java
@@ -1,6 +1,6 @@
-package com.pobluesky.backend.domain.voc.repository;
+package com.pobluesky.backend.domain.answer.repository;
 
-import com.pobluesky.backend.domain.voc.entity.Answer;
+import com.pobluesky.backend.domain.answer.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
