+++ backend/src/main/java/com/pobluesky/backend/domain/question/controller/QuestionController.java
@@ -1,0 +1,28 @@
+package com.pobluesky.backend.domain.question.controller;
+
+import com.pobluesky.backend.domain.question.dto.request.QuestionCreateRequestDTO;
+import com.pobluesky.backend.domain.question.dto.response.QuestionResponseDTO;
+import com.pobluesky.backend.domain.question.service.QuestionService;
+import com.pobluesky.backend.global.util.ResponseFactory;
+import com.pobluesky.backend.global.util.model.JsonResult;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/question")
+public class QuestionController {
+    private final QuestionService questionService;
+
+    @PostMapping("/question/{inquiryId}")
+    public ResponseEntity<JsonResult> createQuestion(
+        @PathVariable Long inquiryId,
+        @RequestBody QuestionCreateRequestDTO questionCreateRequestDTO) {
+        QuestionResponseDTO response = questionService.createQuestion(inquiryId, questionCreateRequestDTO);
+        return ResponseEntity.status(HttpStatus.OK)
+
+            .body(ResponseFactory.getSuccessJsonResult(response));
+    }
+}
