+++ backend/src/main/java/com/pobluesky/backend/domain/question/dto/response/QuestionResponseDTO.java
@@ -1,11 +1,12 @@
-package com.pobluesky.backend.domain.voc.dto.response;
+package com.pobluesky.backend.domain.question.dto.response;
 
-import com.pobluesky.backend.domain.voc.entity.Question;
-import com.pobluesky.backend.domain.voc.entity.QuestionStatus;
+import com.pobluesky.backend.domain.question.entity.Question;
+import com.pobluesky.backend.domain.question.entity.QuestionStatus;
+
 import lombok.Builder;
 
 @Builder
-public record VocQuestionResponseDTO(
+public record QuestionResponseDTO(
     Long inquiryId,
     Long customerId,
     String title,
@@ -13,8 +14,8 @@ public record VocQuestionResponseDTO(
     String files,
     QuestionStatus status
 ) {
-    public static VocQuestionResponseDTO from(Question question) {
-        return VocQuestionResponseDTO.builder()
+    public static com.pobluesky.backend.domain.question.dto.response.QuestionResponseDTO from(Question question) {
+        return com.pobluesky.backend.domain.question.dto.response.QuestionResponseDTO.builder()
             .inquiryId(question.getInquiry().getInquiryId())
             .customerId(question.getInquiry().getCustomer().getCustomerId())
             .title(question.getTitle())
