+++ backend/src/main/java/com/pobluesky/backend/domain/question/repository/QuestionRepository.java
@@ -1,10 +1,13 @@
 package com.pobluesky.backend.domain.question.repository;
 
+import com.pobluesky.backend.domain.inquiry.entity.Inquiry;
 import com.pobluesky.backend.domain.question.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface QuestionRepository extends JpaRepository<Question, Long> {
-
+    Optional<Question> findByInquiry(Inquiry inquiry); // GET 조회 기능 구현 필요
 }
