+++ backend/src/main/java/com/pobluesky/backend/domain/question/service/QuestionService.java
@@ -1,0 +1,39 @@
+package com.pobluesky.backend.domain.question.service;
+
+import com.pobluesky.backend.domain.question.entity.Question;
+import com.pobluesky.backend.domain.answer.entity.Answer;
+import com.pobluesky.backend.domain.inquiry.entity.Inquiry;
+import com.pobluesky.backend.domain.inquiry.repository.InquiryRepository;
+import com.pobluesky.backend.domain.user.repository.CustomerRepository;
+import com.pobluesky.backend.domain.question.dto.request.QuestionCreateRequestDTO;
+import com.pobluesky.backend.domain.question.dto.response.QuestionResponseDTO;
+import com.pobluesky.backend.domain.question.repository.QuestionRepository;
+import com.pobluesky.backend.domain.answer.dto.request.AnswerCreateRequestDTO;
+import com.pobluesky.backend.domain.answer.dto.response.AnswerResponseDTO;
+import com.pobluesky.backend.domain.answer.repository.AnswerRepository;
+import com.pobluesky.backend.global.error.CommonException;
+import com.pobluesky.backend.global.error.ErrorCode;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class QuestionService {
+    private final QuestionRepository questionRepository;
+    private final AnswerRepository answerRepository;
+    private final InquiryRepository inquiryRepository;
+    private final CustomerRepository customerRepository;
+
+    @Transactional
+    public QuestionResponseDTO createQuestion(Long inquiryId, QuestionCreateRequestDTO dto) {
+        Inquiry inquiry = inquiryRepository
+            .findById(inquiryId)
+            .orElseThrow(() -> new CommonException(ErrorCode.INQUIRY_NOT_FOUND));
+
+        Long customerId = inquiry.getCustomer().getCustomerId();
+        Question question = dto.toQuestionEntity(inquiry);
+        Question savedQuestion = questionRepository.save(question);
+        return QuestionResponseDTO.from(savedQuestion);
+    }
+}
