--- backend/src/main/java/com/pobluesky/backend/domain/voc/entity/QuestionStatus.java
@@ -1,15 +1,0 @@
-package com.pobluesky.backend.domain.voc.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum QuestionStatus {
-
-    READY("ready"),
-    COMPLETED("completed");
-
-    private final String status;
-    QuestionStatus(String status) {
-        this.status = status;
-    }
-}
